+++ src/components/FormComponents/FormCheckBox.tsx
@@ -1,6 +1,6 @@
 import { Checkbox, FormControl, FormControlLabel, Typography } from '@material-ui/core'
 import React from 'react'
-import { FormComponentProps } from '../../types';
+import { FormComponentProps } from '../../types/types';
 
 const FormCheckBox: React.FC<FormComponentProps> = ({ question }) => {
   const { name } = question
