+++ src/components/FormComponents/FormText.tsx
@@ -1,6 +1,6 @@
 import { Input, Typography } from '@material-ui/core'
 import React from 'react'
-import { FormComponentProps } from '../../types'
+import { FormComponentProps } from '../../types/types'
 
 const FormText: React.FC<FormComponentProps> = ({ question }) => {
   const { name } = question
